Bugfix Unit Testing with controllers

+++ src/controllers/note.controller.js
@@ -2,10 +2,10 @@ const Note = require('../models/note.model');
 
 
 exports.create = async (noteObj) => {
+    if (!noteObj || !noteObj.title || !noteObj.content) {
+        throw new Error('Title and content are required');
+    }
     try {
-        if (!noteObj.title || !noteObj.content) {
-            throw new Error('Title and content are required');
-        }
         const { title, content } = noteObj;
         let note = new Note({
             title,
@@ -29,6 +29,7 @@ exports.findAll = async () => {
 
 // get a single note
 exports.findOne = async (id) => {
+    if (!id) throw new Error('Note id is required');
     try {
         return await Note.findById(id);
     } catch (error) {
@@ -38,6 +39,10 @@ exports.findOne = async (id) => {
 
 // update a note
 exports.update = async (id, noteObj) => {
+    if (!id) throw new Error('Note id is required');
+    if (!noteObj || !noteObj.title || !noteObj.content) {
+        throw new Error('Title and content are required');
+    }
     try {
         return await Note.findByIdAndUpdate
         (id, noteObj, { new: true });
@@ -49,6 +54,7 @@ exports.update = async (id, noteObj) => {
 
 // delete a note
 exports.delete = async (id) => {
+    if (!id) throw new Error('Note id is required');
     try {
         return await Note.findByIdAndDelete(id);
     } catch (error) {
